Some progress.

+++ lib/dht_config.h
@@ -11,7 +11,7 @@
 /** The size of the triple (hash, IPv4, port). */
 #define DHT_TRIPLE_SIZE (DHT_HASH_SIZE + 4 + 2)
 /** The max. number of triples in a response. */
-#define DHT_MAX_TRIPLES int((DHT_MAX_MESSAGE_SIZE-DHT_HASH_SIZE-1)/DHT_TRIPLE_SIZE)
+#define DHT_MAX_TRIPLES ((DHT_MAX_MESSAGE_SIZE-DHT_HASH_SIZE-1)/DHT_TRIPLE_SIZE)
 /** The max. data response. */
 #define DHT_MAX_DATA_SIZE (DHT_MAX_MESSAGE_SIZE-DHT_HASH_SIZE)
 /** The max. public key size for a START_STREAM message. */
